fix: History Item UI
