+++ mifospay/src/main/java/org/mifos/mobilewallet/mifospay/history/ui/TransactionDetailDialog.kt
@@ -29,7 +29,7 @@ import org.mifos.mobilewallet.mifospay.history.presenter.TransactionDetailPresen
 import org.mifos.mobilewallet.mifospay.receipt.ui.ReceiptActivity
 import org.mifos.mobilewallet.mifospay.common.Constants
 import org.mifos.mobilewallet.mifospay.utils.Toaster
-import org.mifos.mobilewallet.mifospay.utils.Utils.getFormattedAccountBalance
+import org.mifos.mobilewallet.mifospay.common.Utils.getFormattedAccountBalance
 import javax.inject.Inject
 
 /**
