+++ mifospay/src/main/java/org/mifos/mobilewallet/mifospay/history/ui/adapter/SpecificTransactionsAdapter.kt
@@ -16,7 +16,7 @@ import com.mifos.mobilewallet.model.domain.Transaction
 import com.mifos.mobilewallet.model.domain.TransactionType
 import org.mifos.mobilewallet.mifospay.R
 import org.mifos.mobilewallet.mifospay.common.Constants
-import org.mifos.mobilewallet.mifospay.utils.Utils.getFormattedAccountBalance
+import org.mifos.mobilewallet.mifospay.common.Utils.getFormattedAccountBalance
 import javax.inject.Inject
 
 /**
